+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Features:
     - tolower, toupper extended to manage string array.
     - strtrim extended to manage string array.
     - strlength extended to manage string array.
+    - count extended to manage string array.    
     - strcmp, strcmpi, strncmp, strncmpi extended to manage string array.
     - fileread, filewrite extended to manage string array.
     - operators ==, ~=, <, >, >=, <= extended to manage string array.